renamed preprocess_files to prepreocess_files_light and added new one with functions load_subset_data and balance_data and a load_relevant_data_subset. Better docstrings may follow. balancing also reduces size, could be improved by checking first with subset.
